\name{CellCODE-package}
\alias{CellCODE-package}
\alias{CellCODE}
\docType{package}
\title{
Cell type COmputational Differential Estimation
}
\description{
A set of tools for analyzing differntial expression in mixture datasets
}
\details{
\tabular{ll}{
Package: \tab CellCODE\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-08-14\cr
License: \tab What license is it under?\cr
}

}
\author{
Maria Chikina
Maintainer: Complain to <mchikina@gmail.com>
}



