\name{cordist}
\alias{cordist}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Correlation based distance function.
}
\description{
Similar to dist but uses 1-rho as a distance metric.
}
\usage{
cordist(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
A matrix of data.
}
}

\value{
Distance matrix.
}
