\name{Garvan}
\alias{Garvan}
\docType{data}
\title{
Garvan pure blood cell expression dataset.
}
\description{
A dataset of expression profiles for white blood cells.
}
\usage{data("Garvan")}
\format{
  A data frame with 18905 observations on the following 5 variables.
  \describe{
    \item{\code{NEUT}}{a numeric vector}
    \item{\code{TCEL}}{a numeric vector}
    \item{\code{MONO}}{a numeric vector}
    \item{\code{DEND}}{a numeric vector}
    \item{\code{BCEL}}{a numeric vector}
  }
}

\source{
GSE3982
}
