\name{GSE20300}
\alias{GSE20300}
\docType{data}
\title{
Gene Expression data for GSE20300 (Shen-Orr et al. kidney transplant study)
}
\description{
This dataset was made by processing raw data from GEO with RMA. It has many more genes than the data supplied in the supplement of the original paper.
}
