\name{randomizeExp}
\alias{randomizeExp}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
randomise expression
}
\description{
Randomizes linear space data with multiplicative noise resulting in constant variance in log space.
}
\usage{
randomizeExp(data, sd = 0.2)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{data}{
The data to be randomized.
}
  \item{sd}{
Standard deviation for the noise.
}
}

\value{
Randomized data.
}
