\name{resid}
\alias{resid}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Quick residual
}
\description{
A simple function to comupte residuals of a regression.
}
\usage{
resid(dat, lab)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{dat}{
A matrix of data.
}
  \item{lab}{
Either a factor or a model matrix.
}
}

\value{
A matrix of residuals with same dimensions as dat.
}
